% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataframe.R
\name{as_dataframe}
\alias{as_dataframe}
\alias{as.dataframe}
\alias{as_dataframe.default}
\alias{as_dataframe.data.frame}
\alias{as_dataframe.dataframe}
\alias{as_dataframe.list}
\alias{as_dataframe.matrix}
\alias{as_dataframe.table}
\alias{is_dataframe}
\alias{is.dataframe}
\title{Deprecated! Convert objects into dataframes (subclassing tibble) and check for it.}
\usage{
as_dataframe(x, ...)

as.dataframe(x, ...)

\method{as_dataframe}{default}(x, tz = "UTC", ...)

\method{as_dataframe}{data.frame}(x, ..., rownames = "rownames")

\method{as_dataframe}{dataframe}(
  x,
  ...,
  rownames = "rownames",
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

\method{as_dataframe}{list}(
  x,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  ...
)

\method{as_dataframe}{matrix}(x, ..., rownames = "rownames")

\method{as_dataframe}{table}(x, n = "n", ...)

is_dataframe(x)

is.dataframe(x)
}
\arguments{
\item{x}{An object to convert to a \code{dataframe}.}

\item{...}{Additional parameters.}

\item{tz}{The time zone. Useful for converting \code{ts} objects with observations
more frequent than daily.}

\item{rownames}{Name of the column that is prepended to the
\code{dataframe} with the original row names (\code{dataframe}s and \code{tibble}s do not
support row names). If \code{NULL}, row names are dropped. The inclusion of the
rownames column is \strong{not} done if row names are trivial, i.e., they equal
the number of the rows in the data frame.}

\item{.name_repair}{Treatment for problematic column names. \code{"check.unique"}
(default value) do not repair names but make sure they are unique. \code{"unique"}
make sure names are unique and non empty. \code{"universal"} make names unique and
syntactic. \code{"minimal"}do not repair or check (just make sure names exist).}

\item{n}{The name for the column containing the number of items, \code{"n"} by
default.}
}
\value{
A \code{dataframe}, which is an S3 object with class
\code{c("dataframe", "tbl_df", "tbl", "data.frame")}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Convert an object into a \code{dataframe} and check for it. A
\code{dataframe} (without dot) is both a \code{data.frame} (with dot, the default
rectangular dataset structure in R) and a \code{tibble}, the tidyverse
equivalence. In fact, \code{dataframe}s behave almost completely like a \code{tibble},
except for a few details explained in the \strong{details} section.
}
\details{
TODO: explain difference between \code{dataframe}s and \code{tibble}s here...
}
\examples{
class(as.dataframe(mtcars))
class(as.dataframe(tibble::tribble(~x, ~y, 1, 2, 3, 4)))
\donttest{
# Any object, like a vector
v1 <- 1:10
is_dataframe(v1)
(df1 <- as_dataframe(v1))
is_dataframe(df1)
# Check names of an existing dataframe
(as_dataframe(df1, .name_repair = "universal"))
# A data.frame with trivial row names
datasets::iris
as_dataframe(datasets::iris)
# A data.frame containing meaningful row names
datasets::mtcars
as_dataframe(datasets::mtcars)
# A list
l1 <- list(x = 1:3, y = rnorm(3))
as_dataframe(l1)
# A matrix with column and row names
(m1 <- matrix(1:9, nrow = 3L, dimnames = list(letters[1:3], LETTERS[1:3])))
as_dataframe(m1)
# A table
set.seed(756)
(t1 <- table(sample(letters[1:5], 50, replace = TRUE)))
as_dataframe(t1)
# compare with the base R function:
as.data.frame(t1)
}
}
\seealso{
\code{\link[=as_tibble]{as_tibble()}}, \code{\link[=as.data.frame]{as.data.frame()}}
}
\author{
Philippe Grosjean \href{mailto:phgrosjean@sciviews.org}{phgrosjean@sciviews.org}
}
\concept{convert objects}
\keyword{utilities}
