% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.io-package.R
\docType{package}
\name{data-package}
\alias{data-package}
\title{data.io - Data Input/Output, Read or Write Data from Files or Datasets in R Packages in Different Formats}
\description{
Read or write data from many different formats (tabular datasets, images, ...) into R
objects.
}
\section{Important functions}{

\itemize{
\item \code{\link[=read]{read()}} is the main function to read data from R packages or files,
\item \code{\link[=labelise]{labelise()}} adds a \code{label}, and possibly a \code{units} attributes to an
object, to be used while pretty printing a table or plot.
}
}

