% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.io-package.R
\docType{package}
\name{data.io-package}
\alias{data.io-package}
\title{Read and Write Data in Different Formats}
\description{
The \{data.io\} package focuses on reading and writing datasets in different
formats in an unified and convenient way. It can deal with labels and units
metadata for variables, translation in different languages, and even
use a sidecar file for preprocessing the dataset automatically. The same
features are also available for a subset of datasets from R packages.
}
\section{Important functions}{

\itemize{
\item \code{\link[=read]{read()}} is the main function to read data from R packages or files,
\item \code{\link[=write]{write()}} is the main function to write data to disk. It is compatible
with \code{\link[base:write]{base::write()}} but provides many more features if you indicate
\verb{type=} or use it like \code{write$type()}.
\item \code{\link[=labelise]{labelise()}} adds a \code{label}, and possibly a \code{units} attributes to an
object, to be used while pretty printing a table or plot.
}
}

