% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\name{data_example}
\alias{data_example}
\title{Get the path to some example datasets in this package}
\usage{
data_example(path)
}
\arguments{
\item{path}{The subpath to a file inside the "extdata" subdirectory of the
"data.io" package.}
}
\value{
The path to the file, or "" if it is not found.
}
\description{
Get the full path to so example datasets included in different
formats in the "data.io" package.
}
\examples{
data_example("iris.csv")
}
\seealso{
\code{\link[=read]{read()}}
}
\author{
Philippe Grosjean \href{mailto:phgrosjean@sciviews.org}{phgrosjean@sciviews.org}
}
\concept{get package directory}
\keyword{utilities}
