% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{data_types}
\alias{data_types}
\title{List recognized file formats (types) for read() and write()}
\usage{
data_types(types_only = FALSE, view = TRUE)
}
\arguments{
\item{types_only}{If \code{TRUE}, only a vector of types is returned, otherwise,
a \code{tibble} with full specifications is provided.}

\item{view}{If \code{TRUE}, the result is "viewed" (displayed in a table in a
separate window, if the user interface allows it, e.g., in RStudio) and
returned invisibly. Otherwise, the results are returned normally.}
}
\value{
An \code{tibble} with \code{types_only = FALSE}, or a character vector.
}
\description{
Display information about data types that can read() and write()
can use, as well as, the original functions that are delegated (see they
respective help pages for more info and to know which additional parameters
can be used in read() and write()).
}
\details{
The function is mainly designed to be used interactively and to
provide information about file types that can be read() or write(). This
cannot be done through a man page because this list is dynamic and other
packages could add or change entries there. With \code{view = FALSE}, the function
can, nevertheless, be also used in a script or a R Markdown/Notebook
document.
}
\examples{
\dontrun{
data_types()
data_types(TRUE)
}
# For non-interactive use, specify view = FALSE
data_types(view = FALSE)
data_types(TRUE, view = FALSE)
}
\seealso{
\code{\link[=read]{read()}}, \code{\link[=write]{write()}}
}
\author{
Philippe Grosjean \href{mailto:phgrosjean@sciviews.org}{phgrosjean@sciviews.org}
}
\concept{list file types that can be read or write}
\keyword{utilities}
