% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{datasets}
\alias{datasets}
\title{Labelised versions of various datasets provided by other packages}
\description{
From \code{datasets}: use \code{read(<item>, package = "datasets")}
\describe{
\item{\code{\link{iris}}}{Edgar Anderson's iris data.}
\item{\code{\link{trees}}}{Black cherry trees measurements.}
}
}
