% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mauna_loa}
\alias{mauna_loa}
\title{Temperature and atmospheric CO2 at Mauna Loa, Hawai}
\format{An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 768 rows and 4 columns.}
\usage{
mauna_loa
}
\description{
Monthly averages of temperatures and CO2 concentrations, maximal and minimal
monthly temperatures at Mauna Loa slope observatory from 1955 to 2018.
}
\details{
Atmospheric CO2 concentration is mole fraction in dry air, micromol/mol,
abbreviated as ppm. Temperatures are in degree celcius.
}
\examples{
plot(mauna_loa)
}
\keyword{datasets}
