% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_path.R
\name{relative_path}
\alias{relative_path}
\title{Calculate path relative to a reference directory}
\usage{
relative_path(file, dir = getwd())
}
\arguments{
\item{file}{A single string with the path to a file or directory to transform
as relative.}

\item{dir}{A single string with the "reference" directory (by default, the
directory provided by \code{\link[=getwd]{getwd()}}.}
}
\value{
A single character string with the relative path, or \code{file}
unmodified if \code{file} is totally unrelated to \code{dir}.
}
\description{
After normalizing both \code{file} and \code{dir}, try to find a common
ancestor directory to build a path for \code{file} relative to \code{dir}.
}
\examples{
relative_path("/Users/me/project/file.txt", "/Users/me/project")
relative_path("/Users/me/project/subdir/file.txt", "/Users/me/project")
relative_path("/Users/me/file.txt", "/Users/me/project")
relative_path("/Users/me/subdir/file.txt", "/Users/me/project")
relative_path("/Users/file.txt", "/Users/me/project")
relative_path("/Users/subdir1/subdir2/file.txt", "/Users/me/project")
relative_path("/Unrelated/file.txt", "/Users/me/project")
\donttest{
relative_path("file.txt", "/Users/me/project")
relative_path("~/file.txt", "/Users/me/project")
relative_path("./file.txt", "/Users/me/project")
relative_path(file.path(getwd(), "data.io", "file.txt"))
}
}
\seealso{
\code{\link[=getwd]{getwd()}}, \code{\link[=normalizePath]{normalizePath()}}
}
\author{
Philippe Grosjean \href{mailto:phgrosjean@sciviews.org}{phgrosjean@sciviews.org}
}
\concept{
relative paths
}
\keyword{utilities}
