% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{set_label}
\alias{set_label}
\alias{cl}
\title{Set label (and units)}
\usage{
set_label(x, label, units = NULL, as.labelled = FALSE, ...)

cl(..., label = NULL, units = NULL, as.labelled = FALSE)
}
\arguments{
\item{x}{An object.}

\item{label}{The character string to set as \code{label} attribute to \code{x}.}

\item{units}{The units (optional) as a character string to set for \code{x}.}

\item{as.labelled}{Should the object be converted as a \code{labelled} S3 object
(no by default)?}

\item{...}{Further arguments: itesm to be concatenated in a vector using
\code{c(...)} for \code{cl()}.}
}
\value{
The \code{x} object plus a \code{label} attribute, and possibly, a \code{units}
attribute.
}
\description{
Set the \code{label}, as well as the \code{units} attributes to an object.
The label can be used for better display as plot axes labels, or as table
headers in pretty-formatted \R outputs. The units are usually associated to
the label in axes labels for plots. \code{cl()} is a shortcut for concatenate
(\code{c()}) and \code{set_label()}.
}
\details{
The same mechanism as the one used in package \strong{Hmisc} is used
here. However, \strong{Hmisc} always add the \strong{labelled} class to an object,
while here, this is optional. Setting this class make the object more nicely
printed, and subsettable without loosing these attributes. But it conflicts
with a class of the same name in package \strong{haven}, used for other purposes.
So, here, the class is \strong{not} set by default.
}
\examples{
# TODO...
}
\seealso{
\code{\link[=label]{label()}}, \code{\link[=units]{units()}}
}
\author{
Philippe Grosjean \href{mailto:phgrosjean@sciviews.org}{phgrosjean@sciviews.org}
}
\concept{
labelling objects
}
\keyword{utilities}
