% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urchin_growth}
\alias{urchin_growth}
\title{Sea urchins growth}
\format{
An object of class \code{data.frame} with 7024 rows and 3 columns.
}
\usage{
urchin_growth
}
\description{
Size at age for a cohort of farmed sea urchins, \emph{Paracentrotus lividus}.
}
\details{
The same cohort of farmed sea urchins being measured at various time
intervals, the observations are not completely independent from each other:
the same individuals are repeatedly measured here. As the sea urchins are
not individually tagged, it is not possible to track them from one
measurement to the other. However, the whole dataset is representative of the
growth, and spreading of growth in a single cohort. Also, mortality could be
derived from the number of measurements made at each time period, since
\strong{all} the individuals still alive are measured (no sub-sampling).
}
\examples{
library(ggplot2)
ggplot(urchin_growth, aes(age, diameter)) +
  geom_jitter(alpha = 0.2) +
  xlab(svBase::label(urchin_growth$age, units = TRUE)) +
  ylab(svBase::label(urchin_growth$diameter, units = TRUE)) +
  ggtitle("Growth of a cohort of sea urchins")
}
\keyword{datasets}
