% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zooplankton}
\alias{zooplankton}
\title{Zooplankton image analysis}
\format{A data frame with 19 variables:
\describe{
\item{\code{ecd}}{The "equivalent circular diameter", the diameter of a
circle with the same area as the particle (in mm).}
\item{\code{area}}{The area of the particle on the image (in mm^2).}
\item{\code{perimeter}}{The perimeter of the particle (in mm).}
\item{\code{feret}}{The Feret diameter, that is, the largest measured
diameter of the particle on the image (mm).}
\item{\code{major}}{The major axis of the ellipsoid matching the particle
(mm).}
\item{\code{minor}}{The minor axis of the same ellipsoid (mm).}
\item{\code{mean}}{The mean value of the gray levels calibrated in optical
density (OD), thus, unitless.}
\item{\code{mode}}{The most frequent gray level in that particle in OD.}
\item{\code{min}}{The most transparent part in OD.}
\item{\code{max}}{The most opaque part in OD.}
\item{\code{std_dev}}{The stadard deviation of the OD distribution inside
the particle.}
\item{\code{range}}{Transparency range as \code{max} - \code{min}.}
\item{\code{size}}{The mean diameter of the particle, as the average of
\code{minor} and \code{major} (mm).}
\item{\code{aspect}}{Aspect ratio of the particle as \code{minor}/\code{major}.}
\item{\code{elongation}}{The \code{area} divided by the area of a circle of the
same \code{perimeter} of the particle.}
\item{\code{compactness}}{sqrt((4/pi) * \code{area}) / \code{major}.}
\item{\code{transparency}}{1 - (\code{ecd} - \code{size}).}
\item{\code{circularity}}{4pi(\code{area} / \code{perimeter}^2).}
\item{\code{density}}{Density integrate by the surface coevered by each gray
level, i.e. O.D., inside the particle.}
\item{\code{class}}{The classification of this particle. 17 classes are made.
Note that \code{Copepods} are \code{Calanoid} + \code{Cyclopoid} + \code{Harpactivoid} +
\code{Poecilostomatoid} and they represent the most abundant zooplankton at sea.
}
}

This is a typical training set used to train a plankton classifier with
machine learning algorithms. Organisms originate from various samples
(different seasons, depth, etc. to take the variability into account).
However, the abundance of the different classes do \strong{not} match abundance
found in each sample, i.e., rare classes are over-represented in this
training set. Take care that several variables are correlated!}
\source{
Grosjean, Ph & K. Denis (2004). Supervised classification of images,
applied to plankton samples using R and ZooImage. Chap.12 of Data Mining
Applications with R. Zhao, Y. & Y. Cen (eds). Elevier. Pp 331-365.
https://doi.org/10.1016/C2012-0-00333-X.
}
\usage{
zooplankton
}
\description{
Various features measured by image analysis with the package \code{zooimage} and
\code{ImageJ} on samples of zooplankton originating from Tulear, Madagascar. The
taxonomic classification is also provided in the \code{class} variable.
}
\examples{
table(zooplankton$class)
library(ggplot2)
ggplot(zooplankton, aes(circularity, transparency, color = class)) +
  geom_point()
}
\keyword{datasets}
