% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Datasets}
\alias{Datasets}
\title{Labelised versions of various datasets provided by 'data' or other packages}
\description{
Use \code{name <- read("data", package = "pkg", lang = "xx")} to read these
datasets together with the metadata (labels, units, comments, ...).
}
