% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Datasets}
\alias{Datasets}
\title{Labelised versions of various datasets provided by 'data.io' or other packages}
\description{
Use \code{name <- read("data", package = "pkg", lang = "xx")} to read these
datasets together with the metadata (labels, units, comments, ...).
}
\details{
From \code{data}:
\describe{
\item{\code{\link{mauna_loa}}}{Temperature and atmospheric CO2 at Mauna
Loa, Hawai. 5 vars x 768 obs. Time series of monthly averages from 1955 to
2018. }
\item{\code{\link{urchin_bio}}}{Sea urchins biometry. 19 vars x 421 obs.
Morphometric variables measured on two populations of sea urchins, incl.
one circular variable (\code{maturity}). }
\item{\code{\link{urchin_growth}}}{Sea urchins growth. 3 vars x 7024 obs.
Size at age for a cohort of sea urchins followed over more than 10 years. }
\item{\code{\link{zooplankton}}}{Zooplankton image analysis. 20 vars x
1262 obs. A training set with 19 measurements made on images of zooplankton
and their respective class as attributed by taxonomists. }
}

From \code{datasets}:
\describe{
\item{\code{\link{anscombe}}}{Anscombe's quartet of 'identical' simple
linear Regressions. 8 vars x 11 obs. Artificial data. }
\item{\code{\link{iris}}}{Edgar Anderson's iris data. 5 vars x 150 obs.
Morphometry of the flowers of three iris species (50 for each species). }
\item{\code{\link{lynx}}}{Annual Canadian lynx trappings 1821–1934. 2 vars
x 114 obs. Long (> 1 century) time series. }
\item{\code{\link{trees}}}{Black cherry trees measurements. 3 vars x 31
obs. Measurement of tree timber of various sizes. }
}

From \code{ggplot2}:
\describe{
\item{\link[ggplot2:diamonds]{ggplot2::diamonds}}{Prices of 50,000 round cut diamonds. 10 vars
x 53940 obs. Price and other attributes of 10,000's of diamonds. }
\item{\link[ggplot2:mpg]{ggplot2::mpg}}{Fuel economy data from 1999 and 2008 for popular
cars. 11 vars x 234 obs. Data are for most popular U.S. market cars only. }
}

From \code{MASS}:
\describe{
\item{\code{\link{crabs}}}{Morphological measurements on Leptograpsus
crabs. 8 vars x 200 obs. Morphological measurements of \emph{Leptograpsus
variegatus} crabs, either blue or orange, males and females. }
\item{\code{\link{geyser}}}{Old Faithful geyser data. 2 vars x 299 obs.
Duration and waiting time for eruptions from August 1 to August 15, 1985. }
}

From \code{nycflights13}:
\describe{
\item{\link[nycflights13:airlines]{nycflights13::airlines}}{Airlines by their carrier codes. 2 vars x 16
obs. }
\item{\link[nycflights13:airports]{nycflights13::airports}}{Various metadata about New York city
airports. 8 vars x 1458 obs. }
\item{\link[nycflights13:flights]{nycflights13::flights}}{On-time data for all flights that departed NYC
(i.e., JFK, LGA or EWR) in 2013. 19 vars x 336776 obs. }
\item{\link[nycflights13:planes]{nycflights13::planes}}{Planes metadata. 9 vars x 3322 obs. }
\item{\link[nycflights13:weather]{nycflights13::weather}}{Hourly meteorological data for JFK, LGA and
EWR. 15 vars x 26130 obs. }
}
}
