% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Datasets}
\alias{Datasets}
\title{Labelised versions of various datasets provided by 'data' or other packages}
\description{
Use \code{name <- read("data", package = "pkg", lang = "xx")} to read these
datasets together with the metadata (labels, units, comments, ...).
}
\details{
From \code{data}:
\describe{
\item{\code{\link{mauna_loa}}}{Temperature and atmospheric CO2 at Mauna
Loa, Hawai. 5 vars x 768 obs. Time series of monthly averages from 1955 to
2018. }
\item{\code{\link{urchin_bio}}}{Sea urchins biometry. 19 vars x 421 obs.
Morphometric variables measured on two populations of sea urchins, incl.
one circular variable (\code{maturity}). }
\item{\code{\link{urchin_growth}}}{Sea urchins growth. 3 vars x 7024 obs.
Size at age for a cohort of sea urchins followed over more than 10 years. }
\item{\code{\link{zooplankton}}}{Zooplankton image analysis. 20 vars x
1262 obs. A training set with 19 measurements made on images of zooplankton
and their respective class as attributed by taxonomists. }
}

From \code{datasets}:
\describe{
\item{\code{\link{anscombe}}}{Anscombe's quartet of ‘identical’ simple
linear Regressions. 8 vars x 11 obs. Artificial data. }
\item{\code{\link{iris}}}{Edgar Anderson's iris data. 5 vars x 150 obs.
Morphometry of the flowers of three iris species (50 for each species). }
\item{\code{\link{lynx}}}{Annual canadian lynx trappings 1821–1934. 2 vars
x 114 obs. Long (> 1 century) time series. }
\item{\code{\link{trees}}}{Black cherry trees measurements. 3 vars x 31
obs. Measurement of tree timber of various sizes. }
}

From \code{MASS}:
\describe{
\item{\code{\link{crabs}}}{Morphological measurements on Leptograpsus
crabs. 8 vars x 200 obs. Morphological measurements of \emph{Leptograpsus
variegatus} crabs, either blue or orange, males and females. }
\item{\code{\link{geyser}}}{Old Faithful geyser data. 2 vars x 299 obs.
Duration and waiting time for eruptions from August 1 to August 15, 1985. }
}
}
