% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\alias{type_from_extension}
\title{Read data in \R in different formats}
\usage{
read(file, type = NULL, header = "#", header.max = 50L, skip = 0L,
  locale = default_locale(), comments = NULL, package = NULL,
  fun = NULL, fun_list = NULL, ...)

type_from_extension(file)
}
\arguments{
\item{file}{The path to the file to read, or the name of the dataset to get
from an \R package (in that case, you \strong{must} provide the \code{package=}
argument).}

\item{type}{The type (format) of data to read.}

\item{header}{The character to use for the header and other comments.}

\item{header.max}{The maximum of lines to consider for the header.}

\item{skip}{The number of lines to skip at the beginning of the file.}

\item{locale}{The encoding of the file.}

\item{comments}{Comments to add in the created object.}

\item{package}{The package where to look for the dataset. If \code{file=} is not
provided, a list of available datasets in the package is displayed.}

\item{fun}{The function to delegate reading of the data. If \code{NULL} (default),
The function is chosen from \code{fun_list}.}

\item{fun_list}{The table with correspondance of the types, read, and write
functions.}

\item{...}{Further arguments passed to the function \code{fun=}.}
}
\value{
An \R object with the data (its class depends on the data being read).
}
\description{
Read and return an \R object from data on disk, from URL, or
from packages.
}
\details{
\code{read()} allows for a unique entry point to read various kinds of
data, but it delegates the actual work to various other functions dispatched
accross several \R packages. See \code{getOption("read_write")}.
}
\examples{
# TODO...
}
\seealso{
\code{\link[=read_csv]{read_csv()}}
}
\author{
Philippe Grosjean \href{mailto:phgrosjean@sciviews.org}{phgrosjean@sciviews.org}
}
\concept{
read and import data
}
\keyword{utilities}
